+++ app/src/main/AndroidManifest.xml
@@ -28,6 +28,11 @@
             android:label="@string/title_activity_chefs"
             android:theme="@style/AppTheme.NoActionBar" >
         </activity>
+        <activity
+            android:name=".RecommendedLayout"
+            android:label="@string/title_activity_recommended_layout"
+            android:theme="@style/AppTheme.NoActionBar" >
+        </activity>
     </application>
 
 </manifest>
